checkpoint

+++ src/main/java/com/restaurant/repositories/CartRepository.java
@@ -3,6 +3,5 @@ package com.restaurant.repositories;
 import com.restaurant.model.Cart;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CartRepository extends CrudRepository<Cart, Integer> {
-
+public interface CartRepository extends CrudRepository<Cart, Long> {
 }
